ai-engine/pyproject.toml
[tool.poetry]
name = "vytrixe-ai-engine"
version = "0.1.0"
description = "AI Content Automation Engine for Vytrixe"
authors = ["Antigravity <antigravity@google.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = "^0.27.0"
google-generativeai = "^0.4.0"
python-dotenv = "^1.0.1"
moviepy = "^1.0.3"
edge-tts = "^6.1.10"
requests = "^2.31.0"
pydantic = "^2.6.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
